LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/rendering/the-details-element/details-focus-steps-expected.txt

First
Finally Third

FAIL Focus flows into second when details is open assert_equals: expected Element node <a id="second" href="#">Second</a> but got Element node <button id="third">Third</button>
PASS Focus flows into third when details is closed
